+++ RNMixpanel/RNMixpanel.h
@@ -12,6 +12,8 @@
 
 + (RNMixpanel *)sharedInstanceWithToken:(NSString *)apiToken
                 optOutTrackingByDefault:(BOOL)optOutTrackingByDefault
+                           trackCrashes:(BOOL)trackCrashes
+                  automaticPushTracking:(BOOL)automaticPushTracking
                           launchOptions:(NSDictionary *)launchOptions;
 
 @end